+++ swing/common-ui/src/main/java/module-info.java
@@ -62,6 +62,7 @@ module is.codion.swing.common.ui {
 	exports is.codion.swing.common.ui.component.indicator;
 	exports is.codion.swing.common.ui.component.value;
 	exports is.codion.swing.common.ui.border;
+	exports is.codion.swing.common.ui.color;
 	exports is.codion.swing.common.ui.control;
 	exports is.codion.swing.common.ui.dialog;
 	exports is.codion.swing.common.ui.frame;