+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTableCellRenderer.java
@@ -31,7 +31,7 @@ import is.codion.swing.framework.model.SwingEntityTableModel;
 
 import java.awt.Color;
 
-import static is.codion.swing.common.ui.Colors.darker;
+import static is.codion.swing.common.ui.color.Colors.darker;
 
 /**
  * @see #builder(Attribute, SwingEntityTableModel)